+++ app/src/main/res/layout/fragment_board_view.xml
@@ -115,13 +115,13 @@
 
     </androidx.core.widget.NestedScrollView>
 
-    <uk.whitecrescent.waqti.frontend.customview.TaskListLayout
+    <uk.whitecrescent.waqti.frontend.customview.draglayouts.TaskListLayout
         android:id="@+id/list_dragView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         tools:ignore="MissingConstraints" />
 
-    <uk.whitecrescent.waqti.frontend.customview.TaskCardLayout
+    <uk.whitecrescent.waqti.frontend.customview.draglayouts.TaskCardLayout
         android:id="@+id/task_dragView"
         android:layout_width="0dp"
         android:layout_height="0dp"
